requirements.txt
aiohttp==3.7.1
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
Flask-Limiter==1.4
Flask-SocketIO==4.3.1
Flask-WTF==0.14.3
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
monotonic==1.5
multidict==5.0.0
packaging==20.4
pip-review==1.1.0
prettytable==1.0.1
pyee==7.0.4
pyparsing==2.4.7
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
slackclient==2.9.3
slackeventsapi==2.2.1
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.6.2
